Improved Date Selection, Autocomplete Functionality, and Intuitive Saved Trips Management
I addressed the issues with the date selection in the v-text-field, ensuring that the calendar now appears when the area is selected. Users can click on a date on the calendar, and the selected date gets saved properly.

The autocomplete section for selecting a city has been fixed. All previously encountered problems with this component have been resolved.

In the saved trips view, I implemented functionality so that when the trash icon is clicked, an alert will prompt the user for confirmation. If the user clicks on "confirm," the corresponding card will be deleted. Additionally, I implemented a system where the city and associated information are stored in an array. The view pulls information from this array to display the saved trips, creating a seamless and dynamic user experience.